+++ packages/external-db-authorization/lib/auth-providers/aws_auth_provider.js
