+++ packages/external-db-authorization/lib/auth-providers/azure_auth_provider.js
