+++ packages/external-db-authorization/lib/auth-providers/gcp_auth_provider.js
