+++ packages/external-db-authorization/lib/auth-providers/stub_auth_provider.js
